refactor(healthkit): apply review feedback — split views, fix error handling, narrow HealthKit platform filters
Agent-Logs-Url: https://github.com/thef4tdaddy/NucleOS/sessions/03a096d0-8dac-46d5-ae3c-f84329078ec9
